real_estate/apps/page/serializers.py
# serializers for the page app 

from rest_framework import serializers
from .models import Page

class PageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Page
        fields = '__all__'
        pass
    