pyproject.toml
[tool.poetry]
name = "real-estate"
version = "0.1.0"
description = ""
authors = ["r0w4nr00t <r0w4nr00t@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
markdown = "^3.7"
django-filter = "^24.3"
pillow = "^11.0.0"
django-phonenumber-field = "^8.0.0"
django-environ = "^0.11.2"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg = "^3.2.3"
mailtrap = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
